FUSE: Trunkierungsstand persistent, zwei Races beseitigt
Tonis vier Befunde vom 18.08.

TRUNKIERUNGSZAEHLER WAR NICHT PERSISTENT. `entfernteErgebnisse` galt nur seit
Prozessstart und fehlte im Codec - nach einem Neustart meldete eine laengst
gekappte Generation wieder `historyTruncated=false`. Gerade mehrtaegige
Messdaten haetten also genau dann vollstaendig ausgesehen, wenn sie es am
wenigsten sind. Der Gesamtstand steht jetzt als `droppedTotal` in DERSELBEN
Generation, die kappt, und wird beim Laden uebernommen.

RACE AN DER LUECKENMARKE. `@Volatile` macht `maxOf(lesen, neu)` nicht
unteilbar: ein Queue-Drop im Loop-Thread und das Laden im Worker konnten sich
gegenseitig ueberschreiben - und ausgerechnet die Lueckenmarke darf nie
zurueckfallen. Jetzt AtomicLong.accumulateAndGet(::maxOf).

RACE AN DER TELEMETRIE. Producer und Worker ersetzten parallel
`telemetry.copy(...)` - ein Lese-Aendern-Schreiben ueber mehrere Felder, bei
dem Schreibdauer, asOfTs oder Drop-Zaehler verlorengingen. Ausgerechnet die
LASTMESSUNG waere damit unzuverlaessig gewesen, also die Zahl, wegen der es
den Schalter gibt. Telemetry ist jetzt eine BERECHNETE SICHT auf atomare
Einzelfelder.

DIE RECORDER-KAPPUNG IST DOCH TESTBAR. Ich hatte sie zu schnell abgetan;
Tonis Bauanleitung funktioniert: Store mit vollem Bestand vorladen, eine
faellige Erwartung dazugeben, einen Zyklus abrechnen lassen. Beim ersten
Anlauf lag die Fuellung ausserhalb von OUTCOME_RETENTION_MS (4 h) und war
beim ersten advance verschwunden - dicht vor den Zeitpunkt gelegt, greift die
Kappung wie erwartet. Drei Mutationsproben beissen jetzt, wo vorher eine
stumpf blieb.

Weiterhin offen: scharfer Runner-Integrationstest DORMANT -> CORRECTION.

+++ fuse/plugin/src/main/kotlin/app/aaps/fuse/plugin/export/FuseStateJson.kt
@@ -1029,7 +1029,7 @@ object FuseStateJson {
         val historyTruncated: Boolean,
         /** Monoton: wie viele Ergebnisse die Kappung seit Prozessstart
          *  entfernt hat. Eine gekappte Strecke ist "mindestens N Minuten". */
-        val droppedOutcomesTotal: Int,
+        val droppedOutcomesTotal: Long,
         /** Aeltestes noch gehaltenes Ergebnis - davor ist nichts mehr da. */
         val oldestRetainedDueTs: Long,
     )
